Create Random_Walk.c
Professor Confunkulus enjoys taking refreshing strolls in the evening, but because he is a mathematics professor, he wishes to make his evening walks as mathematically convoluted as possible. He takes a field and divides it into a square matrix of size 10. To each cell in the matrix, he assigns a random number between 1 and 9. Each number represents a direction, as one would see on a typical keyboard number pad. Professor Confunkulus moves through the field one cell at a time, always following the direction indicated in the cell he’s currently standing in. 

He stops if:
• He is currently standing in a 5-cell.
• His walk would take him outside the field.
• His walk would have him return to a previously visited cell.

Your mission, should you choose to accept it, is to write a C function which takes a 10x10 array of integers and a pointer to a starting location within that array and outputs a pointer to his final position within the array.